79 sprint 15   validação de orçamento

+++ Civitas.WebAPI/Objects/Dtos/Entities/InstituicaoOrcamentoDisponivelDTO.cs
@@ -4,6 +4,6 @@ namespace Civitas.WebAPI.Objects.Dtos.Entities
     {
         public int IdInstituicao { get; set; }
         public string NomeInstituicao { get; set; }
-        public double TotalOrcamentoDisponivel { get; set; }
+        public decimal TotalOrcamentoDisponivel { get; set; }
     }
 }
